Add support for generating embeddings from multi-modal inputs, not just text. Certain Google models support this and the PHP AI Client already supports this so we should support that here
