+++ .gitignore
@@ -1,2 +1,3 @@
 /vendor/
+/.env
 /.phpunit.result.cache
