+++ README.md
@@ -62,10 +62,17 @@ echo $result->toText();
 
 ### Generating Embeddings
 
+Embedding generation always requires you to name a model. Vectors are only comparable to other
+vectors produced by the same model, so a stored corpus is permanently tied to the model that created
+it — letting the library pick could silently invalidate embeddings you already saved.
+
 ```php
+use WordPress\AiClient\AiClient;
+use WordPress\GoogleAiProvider\Provider\GoogleProvider;
+
 // Generate a single embedding.
 $embedding = AiClient::input('PHP powers a large part of the web.')
-    ->usingProvider('google')
+    ->usingModel(GoogleProvider::model('gemini-embedding-001'))
     ->generateEmbedding();
 
 $values = $embedding->getValues();
@@ -75,19 +82,69 @@ $embeddings = AiClient::input([
         'PHP powers a large part of the web.',
         'WordPress makes publishing accessible.',
     ])
-    ->usingProvider('google')
+    ->usingModel(GoogleProvider::model('gemini-embedding-001'))
     ->generateEmbeddings();
 
-// Request a specific number of output dimensions (for models that support it).
+// Request a specific number of output dimensions.
 $embedding = AiClient::input('PHP powers a large part of the web.')
-    ->usingProvider('google')
+    ->usingModel(GoogleProvider::model('gemini-embedding-001'))
     ->usingDimensions(512)
     ->generateEmbedding();
 ```
 
+#### Multimodal Embeddings
+
+The `gemini-embedding-2` model embeds images, audio, video and PDF documents in addition to text,
+mapping every modality into the same vector space. That means a text embedding can be compared
+directly against an image embedding, which is what makes cross-modal search possible.
+
+```php
+// Embed an image. Remote URLs, such as WordPress attachment URLs, are passed through directly.
+$embedding = AiClient::input(new File(wp_get_attachment_url($attachment_id)))
+    ->usingModel(GoogleProvider::model('gemini-embedding-2'))
+    ->generateEmbedding();
+
+// Text and media can be mixed in a single batch. One embedding is returned per input, in order.
+$embeddings = AiClient::input([
+        'A golden retriever running on a beach.',
+        new File('https://example.com/dog-on-beach.jpg'),
+        new File('https://example.com/clip.mp4'),
+    ])
+    ->usingModel(GoogleProvider::model('gemini-embedding-2'))
+    ->generateEmbeddings();
+```
+
+Input modalities are advertised per model, so the model you name is matched against the inputs you
+provide. Passing an image to a text-only model such as `gemini-embedding-001` is rejected. On PHP AI
+Client versions that validate the named model, this happens locally before any request is sent, with
+a message naming the unsupported option:
+
+```
+Model "gemini-embedding-001" from provider "google" cannot fulfill this embedding request.
+Unsupported options: inputModalities ([image]).
+```
+
+Supported input per model:
+
+| Model | Text | Image | Audio | Video | PDF |
+| --- | --- | --- | --- | --- | --- |
+| `gemini-embedding-2` | Yes | Yes | Yes | Yes | Yes |
+| `gemini-embedding-001` | Yes | No | No | No | No |
+
+A few Google-specific notes:
+
+* Each input produces its own embedding. Combining several parts into a single fused vector is not
+  currently expressible through the PHP AI Client.
+* `taskType` is honored by `gemini-embedding-001` but ignored by `gemini-embedding-2`; for the latter,
+  describe the task in the input text instead.
+* `gemini-embedding-2` normalizes truncated dimension vectors automatically, while
+  `gemini-embedding-001` requires callers to renormalize them.
+* `gemini-embedding-2` reports token usage on the embedding result; `gemini-embedding-001` does not,
+  and reports zero.
+
 ## Supported Models
 
-Available models are dynamically discovered from the Google AI API. This includes Gemini models for text generation (with multimodal input support), Imagen models for image generation, and embedding models (such as `gemini-embedding-001` and `text-embedding-004`) for embedding generation. See the [Google AI documentation](https://ai.google.dev/gemini-api/docs/models) for the full list of available models.
+Available models are dynamically discovered from the Google AI API. This includes Gemini models for text generation (with multimodal input support), Imagen models for image generation, and embedding models (such as `gemini-embedding-2` for multimodal embeddings and `gemini-embedding-001` for text-only embeddings). See the [Google AI documentation](https://ai.google.dev/gemini-api/docs/models) for the full list of available models.
 
 ## Configuration
